% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getconfuMat.R
\name{getconfuMat}
\alias{getconfuMat}
\title{MegaR analysis}
\usage{
getconfuMat(testdata, rfmodel)
}
\arguments{
\item{testdata}{testdata}

\item{rfmodel}{the model on which classification is done}
}
\description{
This is an internal function used to plot the confusion Matrix
}
