% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoodfeature.R
\name{getGoodfeature}
\alias{getGoodfeature}
\title{MegaR getGoodfeature}
\usage{
getGoodfeature(alltable2, threshold, samplePercent, normval)
}
\arguments{
\item{alltable2}{table containing all the features analyzed}

\item{threshold}{threshold of the value of feature that should
be across the sample}

\item{samplePercent}{percentage of sample that should contain
the threshold amount of value in its feature}

\item{normval}{wheather cumulative sum normalization is to be used or not}
}
\description{
This is an internal function used to collect good features
}
