% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevelData.R
\name{getLevelData}
\alias{getLevelData}
\title{MegaR getLevelData}
\usage{
getLevelData(alltable, leveld)
}
\arguments{
\item{alltable}{the taxonomic table}

\item{leveld}{the taxonomic level at which to select the feature}
}
\description{
This is an internal function used to extract either the species or genus
level of information
}
