% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettrainingdonesvm.R
\name{gettrainingdonesvm}
\alias{gettrainingdonesvm}
\title{MegaR gettrainingdonesvm}
\usage{
gettrainingdonesvm(mytable3, classid, sampleid, ruleout, psd, metadat,
  svmmethod)
}
\arguments{
\item{mytable3}{processed input file with features}

\item{classid}{the column number in metadata file in which the class of
input data is stored}

\item{sampleid}{the column number of metadata file which contain sample ids
that match with input data}

\item{ruleout}{the class which is to be removed from classification model}

\item{psd}{the percentage of data to be split into training set}

\item{metadat}{the metadata path}

\item{svmmethod}{one of the many svm method available in caret}
}
\description{
This is the function to get class information for the input data from the
metadata file and build the support vector machines as predictive models.
}
