% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getunknpredict.R
\name{getunknpredict}
\alias{getunknpredict}
\title{MegaR analysis}
\usage{
getunknpredict(unknormdata, a)
}
\arguments{
\item{unknormdata}{unknown dataset}

\item{a}{model list with elements ******}
}
\description{
This is an internal function used to predict the given dataset.
}
