% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaRshiny.R
\name{megaR}
\alias{megaR}
\title{megaR megaRshiny
Use megaR through shiny interface}
\usage{
megaR()
}
\description{
This function allows the user to input data files and alter the input
variables to make sure the formatting is correct.
They can then run the megaR package which will output the results and plots
in the browser and allow the user to download results as needed.
}
\examples{
if(interactive()) {LONGO()}
}
