% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimportantfeatures.R
\name{plotimportantfeatures}
\alias{plotimportantfeatures}
\title{This function plots the top 10 important features for random forest
classification.}
\usage{
plotimportantfeatures(RF_state_classify, noOffeature)
}
\arguments{
\item{RF_state_classify}{the random forest model}

\item{noOffeature}{no. of feature =10}
}
\description{
This function plots the top 10 important features for random forest
classification.
}
