% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotimptfeatures.R
\name{plotimptfeatures}
\alias{plotimptfeatures}
\title{MegaR analysis
This function plots the top 10 important features for random forest
classification.}
\usage{
plotimptfeatures(RF_state_classify, noOffeature)
}
\arguments{
\item{RF_state_classify}{the random forest model}

\item{noOffeature}{no. of feature =10}
}
\description{
MegaR analysis
This function plots the top 10 important features for random forest
classification.
}
