% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmetadata.R
\name{readmetadata}
\alias{readmetadata}
\title{megaR analysis
This is internal function to read data}
\usage{
readmetadata(x)
}
\arguments{
\item{x}{the path to the file}
}
\description{
megaR analysis
This is internal function to read data
}
