% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{MegaR savePlot}
\usage{
savePlot(file, plotIn)
}
\arguments{
\item{file}{name of the file where data is to be stored}

\item{plotIn}{the plot which is stored
import ggsave from ggplot2**}
}
\description{
This is a internal function. This function allows the program to
save the plot generated in the program.
}
