% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{megaR validation
This function conducts 10 fold cross validation on N set of data}
\usage{
validation(N, modelclas, mytable3, classid, sampleid, ruleout, psd,
  metadat)
}
\arguments{
\item{modelclas}{one of the model, randomforest, supportvector machine or 
generalized linear model}

\item{mytable3}{processed input file with features}

\item{classid}{the column number in metadata file in which the class of 
input data is stored}

\item{sampleid}{the column number of metadata file which contain sample ids
that match with input data
@param psd the percentage of data to be split into training set
@param metadat the metadata path
@param svmmethod one of the many svm method available in caret

@export}

\item{ruleout}{the class which is to be removed from classification model}
}
\description{
megaR validation
This function conducts 10 fold cross validation on N set of data
}
