% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{MegaR validation}
\usage{
validation(Num, modelclas, mytable3, classid, sampleid, ruleout, psd,
  metadat)
}
\arguments{
\item{Num}{No. of sets to run for validation}

\item{modelclas}{one of the model, randomforest, supportvector machine or
generalized linear model}

\item{mytable3}{processed input file with features}

\item{classid}{the column number in metadata file in which the class of
input data is stored}

\item{sampleid}{the column number of metadata file which contain sample ids
that match with input data}

\item{ruleout}{the class which is to be removed from classification model}

\item{psd}{the percentage of data to be split into training set}

\item{metadat}{the metadata path}
}
\description{
This function conducts 10 fold cross validation on N set of data
}
